Extract terminal instance hook
